[추가/수정] totalpage 계산식 수정

+++ src/main/java/kr/co/grib/drools/api/CRules/dto/CRuleListResponseCtlDto.kt
@@ -21,5 +21,5 @@ class CRuleListPagingDto(
     @Schema(description = "totalCount")
     var totalCount: Int ?= null,
     @Schema(description = "totalPages")
-    var totalPages: Long = 0,
+    var totalPages: Int = 0,
 )
